+++ package.json
@@ -1,52 +1,24 @@
 {
     "name": "opentrackmania",
-    "version": "1.0.0",
+    "version": "3.0.0",
     "description": "",
-    "main": "index.js",
+    "engines": {
+      "node": "10.x"
+    },
     "scripts": {
-        "update:seasons": "ts-node updateSeasons.ts",
-        "update:leaderboard": "ts-node updateLeaderboard.ts",
-        "update:totd": "ts-node updateTOTD.ts",
-        "start": "ts-node index.ts",
-        "heroku-postbuild": "cd client && npm i && npm i --only=dev --no-shrinkwrap && npm run build",
-        "watch": "tsnd --respawn index.ts",
-        "tslint": "tslint -c tslint.json -p tsconfig.json",
-        "migrate": "sequelize db:migrate",
-        "migrate:undo": "sequelize db:migrate:undo",
-        "seed:all": "sequelize db:seed:all",
-        "seed:undo:all": "sequelize db:seed:undo:all"
+      "start": "cd client/ && npm i && npm run build && cd ../server/ && npm i && npm run start"
     },
     "cacheDirectories": [
-        "node_modules",
-        "react-ui/node_modules"
+      "server/node_modules",
+      "client/node_modules"
     ],
-    "keywords": [],
-    "author": "",
-    "license": "ISC",
-    "dependencies": {
-        "@types/express": "^4.17.7",
-        "@types/node": "^14.0.27",
-        "@types/node-cron": "^2.0.3",
-        "dotenv": "^8.2.0",
-        "express": "^4.17.1",
-        "node-cache": "^5.1.2",
-        "node-cron": "^2.0.3",
-        "pg": "^8.3.0",
-        "pg-hstore": "^2.3.3",
-        "sequelize": "^6.3.3",
-        "trackmania-api-node": "0.0.22",
-        "ts-node": "^8.10.2",
-        "typescript": "^3.9.7"
+    "repository": {
+      "type": "git",
+      "url": "https://github.com/breeku/opentrackmania.git"
     },
-    "devDependencies": {
-        "@typescript-eslint/eslint-plugin": "^3.8.0",
-        "@typescript-eslint/parser": "^3.8.0",
-        "dotenv-cli": "^3.2.0",
-        "eslint": "^6.6.0",
-        "eslint-config-prettier": "^6.11.0",
-        "eslint-plugin-prettier": "^3.1.4",
-        "prettier": "^2.0.5",
-        "sequelize-cli": "^6.2.0",
-        "ts-node-dev": "^1.0.0-pre.56"
-    }
-}
+    "keywords": [
+      "trackmania"
+    ],
+    "license": "MIT",
+    "devDependencies": {}
+  }
