+++ server/.eslintrc.js
