+++ server/tsconfig.json
@@ -66,6 +66,6 @@
         "skipLibCheck": true /* Skip type checking of declaration files. */,
         "forceConsistentCasingInFileNames": true /* Disallow inconsistently-cased references to the same file. */
     },
-    "include": ["server/**/*.ts"],
+    "include": ["./**/*.ts"],
     "exclude": ["node_modules/**", "client/**"]
 }
